refactor(core): move page_models, backup_schema, rulebook to core package (PR3)
- Move page_models.py -> core/page_models.py
- Move backup_schema.py -> core/backup_schema.py
- Move rulebook.py -> core/rulebook.py
- Update all imports in services/, pages/, data/, tests/
- Delete old top-level module files
