+++ src/handoff/data/queries.py
@@ -11,9 +11,9 @@ from sqlalchemy.orm import selectinload
 from sqlmodel import or_, select
 
 from handoff.core.models import CheckIn, CheckInType, Handoff, Project
+from handoff.core.page_models import HandoffQuery
 from handoff.data.handoffs import _latest_check_in, handoff_is_open
 from handoff.db import session_context
-from handoff.page_models import HandoffQuery
 
 
 def _last_check_in_is_delayed(handoff: Handoff) -> bool:
