+++ src/handoff/pages/now.py
@@ -12,9 +12,9 @@ from datetime import date
 import streamlit as st
 
 from handoff.core.models import CheckIn, CheckInType, Handoff, Project
+from handoff.core.rulebook import BuiltInSection
 from handoff.dates import add_business_days, format_date_smart, format_risk_reason
 from handoff.instrumentation import time_action
-from handoff.rulebook import BuiltInSection
 from handoff.services import (
     add_check_in,
     conclude_handoff,
