+++ src/handoff/pages/projects.py
@@ -9,7 +9,7 @@ import streamlit as st
 from loguru import logger
 
 from handoff.autosave import autosave_editor
-from handoff.page_models import ProjectSummaryRow
+from handoff.core.page_models import ProjectSummaryRow
 from handoff.services.project_service import (
     archive_project,
     create_project,
