+++ src/handoff/services/settings_service.py
@@ -6,15 +6,15 @@ import json
 from pathlib import Path
 from typing import Any
 
-from handoff.data import get_export_payload as _get_export_payload
-from handoff.data import import_payload as _import_payload
-from handoff.rulebook import (
+from handoff.core.rulebook import (
     DEFAULT_RISK_RULE_ID,
     DeadlineWithinDaysCondition,
     RulebookSettings,
     RuleDefinition,
     build_default_rulebook_settings,
 )
+from handoff.data import get_export_payload as _get_export_payload
+from handoff.data import import_payload as _import_payload
 
 DEFAULT_DEADLINE_NEAR_DAYS = 1
 DEADLINE_NEAR_DAYS_MIN = 1
